# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MateScope",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/two-factor/verify": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Verify",
        "operationId": "verify_api_v1_auth_two_factor_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FactorProof"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/two-factor/enroll": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Enroll",
        "operationId": "enroll_api_v1_auth_two_factor_enroll_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/two-factor/confirm": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Confirm",
        "operationId": "confirm_api_v1_auth_two_factor_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryCodesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/two-factor/disable": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Disable",
        "operationId": "disable_api_v1_auth_two_factor_disable_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManagementInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/two-factor/recovery-codes": {
      "post": {
        "tags": [
          "authentication"
        ],
        "summary": "Regenerate",
        "operationId": "regenerate_api_v1_auth_two_factor_recovery_codes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ManagementInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryCodesResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AuthResponse": {
        "properties": {
          "csrf_token": {
            "type": "string",
            "title": "Csrf Token"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "status": {
            "type": "string",
            "const": "authenticated",
            "title": "Status",
            "default": "authenticated"
          }
        },
        "type": "object",
        "required": [
          "csrf_token",
          "username"
        ],
        "title": "AuthResponse"
      },
...
      "ConfirmInput": {
        "properties": {
          "current_password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "format": "password",
            "title": "Current Password",
            "writeOnly": true
          },
          "code": {
            "type": "string",
            "maxLength": 6,
            "minLength": 6,
            "format": "password",
            "title": "Code",
            "writeOnly": true
          }
        },
        "type": "object",
        "required": [
          "current_password",
          "code"
        ],
        "title": "ConfirmInput"
      },
...
      "EnrollmentResponse": {
        "properties": {
          "secret": {
            "type": "string",
            "title": "Secret"
          },
          "provisioning_uri": {
            "type": "string",
            "title": "Provisioning Uri"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In",
            "default": 600
          }
        },
        "type": "object",
        "required": [
          "secret",
          "provisioning_uri"
        ],
        "title": "EnrollmentResponse"
      },
      "FactorProof": {
        "properties": {
          "method": {
            "type": "string",
            "enum": [
              "totp",
              "recovery_code"
            ],
            "title": "Method"
          },
          "code": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "format": "password",
            "title": "Code",
            "writeOnly": true
          }
        },
        "type": "object",
        "required": [
          "method",
          "code"
        ],
        "title": "FactorProof"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ManagementInput": {
        "properties": {
          "current_password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "format": "password",
            "title": "Current Password",
            "writeOnly": true
          },
          "proof": {
            "$ref": "#/components/schemas/FactorProof"
          }
        },
        "type": "object",
        "required": [
          "current_password",
          "proof"
        ],
        "title": "ManagementInput"
      },
...
      "PasswordInput": {
        "properties": {
          "current_password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "format": "password",
            "title": "Current Password",
            "writeOnly": true
          }
        },
        "type": "object",
        "required": [
          "current_password"
        ],
        "title": "PasswordInput"
      },
...
      "RecoveryCodesResponse": {
        "properties": {
          "csrf_token": {
            "type": "string",
            "title": "Csrf Token"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "status": {
            "type": "string",
            "const": "authenticated",
            "title": "Status",
            "default": "authenticated"
          },
          "recovery_codes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recovery Codes"
          }
        },
        "type": "object",
        "required": [
          "csrf_token",
          "username",
          "recovery_codes"
        ],
        "title": "RecoveryCodesResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}